requirement.txt
fastapi
uvicorn[standard]

# Database
motor
pymongo

# Validation
pydantic
email-validator

# Security
passlib
argon2-cffi
python-jose[cryptography]

# Environment
python-dotenv

